# openapi.yaml
openapi: 3.1.0
info:
  title: Gatekeeper
  description: >
    API gateway on Cloudflare Workers with AWS IAM-style policy authorization.
    Fronts Cloudflare's cache purge API and S3-compatible R2 storage with
    per-key/credential policies, token-bucket rate limiting, request collapsing,
    and D1-backed analytics. The S3 proxy provides fine-grained object-level
    access control that R2's native tokens cannot — key-prefix restrictions,
    per-operation permissions, cross-bucket mixed policies.
  version: 3.0.0
  contact:
    name: Erfi Anugrah
servers:
  - url: https://gate.erfi.io
    description: Production

# ─── Security schemes ──────────────────────────────────────────────────────
security: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: >
        Gateway API key (prefix `gw_`). Issued via `POST /admin/keys`.
        Each key has an attached IAM policy that controls which actions,
        resources, and conditions it can access.

    AdminKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: >
        Shared admin secret for CLI and automation. Falls back when
        Cloudflare Access is not configured.

    CloudflareAccess:
      type: apiKey
      in: header
      name: Cf-Access-Jwt-Assertion
      description: >
        Cloudflare Access JWT. Injected automatically by Access for
        browser-based dashboard users. Also accepted via `CF_Authorization`
        cookie.

    S3SigV4Auth:
      type: apiKey
      in: header
      name: Authorization
      description: >
        AWS Signature Version 4 authentication. S3 clients (rclone, boto3,
        aws-cli) send this automatically. Credentials are issued via
        `POST /admin/s3/credentials` (access key prefix `GK`). Also
        supports presigned URL authentication via query parameters.

  # ─── Reusable schemas ──────────────────────────────────────────────────
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [success, errors]
      properties:
        success:
          type: boolean
          const: false
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
        message:
          type: string

    # --- Bulk operation results ---
    BulkItemResult:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
        status:
          type: string
          enum: [revoked, deleted, already_revoked, not_found]
    BulkResult:
      type: object
      required: [processed, results]
      properties:
        processed:
          type: integer
          description: Number of items processed.
        results:
          type: array
          items:
            $ref: '#/components/schemas/BulkItemResult'
    BulkDryRunItem:
      type: object
      required: [id, current_status, would_become]
      properties:
        id:
          type: string
        current_status:
          type: string
          enum: [active, revoked, expired, not_found]
        would_become:
          type: string
    BulkDryRunResult:
      type: object
      required: [dry_run, would_process, items]
      properties:
        dry_run:
          type: boolean
          const: true
        would_process:
          type: integer
          description: Number of items that would be processed.
        items:
          type: array
          items:
            $ref: '#/components/schemas/BulkDryRunItem'

    # --- Policy document ---
...
    UpstreamR2:
      type: object
      required: [id, name, bucket_names, access_key_preview, endpoint, created_at, revoked]
      properties:
        id:
          type: string
          description: Endpoint ID (prefix `upr2_`).
          examples: ['upr2_a1b2c3d4']
        name:
          type: string
        bucket_names:
          type: string
          description: Comma-separated bucket names, or `*` for all buckets.
        access_key_preview:
          type: string
          description: First 8 characters of the access key ID for identification.
        endpoint:
          type: string
          description: R2 S3-compatible endpoint URL.
        created_at:
          type: integer
          description: Unix timestamp (ms).
        created_by:
          type: ['string', 'null']
          description: Email of the creator (from Access JWT), or null.
        revoked:
          type: integer
          enum: [0, 1]
    CreateUpstreamR2Request:
      type: object
      required: [name, access_key_id, secret_access_key, endpoint, bucket_names]
      properties:
        name:
          type: string
        access_key_id:
          type: string
          description: R2 access key ID.
        secret_access_key:
          type: string
          description: R2 secret access key.
        endpoint:
          type: string
          description: R2 S3-compatible endpoint URL.
        bucket_names:
          type: array
          minItems: 1
          items:
            type: string
          description: >
            Bucket names this endpoint serves, or `["*"]` for all buckets.
        created_by:
          type: string

    # --- Config ---
...
paths:
...
  /admin/upstream-r2:
    post:
      operationId: createUpstreamR2
      summary: Register upstream R2 endpoint
      description: >
        Register R2 credentials for proxying S3 requests. Maps bucket names
        to an R2 endpoint with access key and secret. Credentials are
        encrypted at rest.
      tags: [UpstreamR2]
      security:
        - CloudflareAccess: []
        - AdminKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUpstreamR2Request'
      responses:
        '200':
          description: R2 endpoint registered.
          content:
            application/json:
              schema:
                type: object
                required: [success, result]
                properties:
                  success:
                    type: boolean
                    const: true
                  result:
                    $ref: '#/components/schemas/UpstreamR2'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

    get:
      operationId: listUpstreamR2
      summary: List upstream R2 endpoints
      tags: [UpstreamR2]
      security:
        - CloudflareAccess: []
        - AdminKeyAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, revoked]
      responses:
        '200':
          description: List of R2 endpoints (secrets redacted).
          content:
            application/json:
              schema:
                type: object
                required: [success, result]
                properties:
                  success:
                    type: boolean
                    const: true
                  result:
                    type: array
                    items:
                      $ref: '#/components/schemas/UpstreamR2'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /admin/upstream-r2/{id}:
    get:
      operationId: getUpstreamR2
      summary: Get upstream R2 endpoint details
      tags: [UpstreamR2]
      security:
        - CloudflareAccess: []
        - AdminKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Upstream R2 endpoint ID (prefix `upr2_`).
      responses:
        '200':
          description: Endpoint details (secrets redacted).
          content:
            application/json:
              schema:
                type: object
                required: [success, result]
                properties:
                  success:
                    type: boolean
                    const: true
                  result:
                    $ref: '#/components/schemas/UpstreamR2'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Endpoint not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

    delete:
      operationId: revokeUpstreamR2
      summary: Revoke upstream R2 endpoint
      tags: [UpstreamR2]
      security:
        - CloudflareAccess: []
        - AdminKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Endpoint revoked.
          content:
            application/json:
              schema:
                type: object
                required: [success, result]
                properties:
                  success:
                    type: boolean
                    const: true
                  result:
                    type: object
                    required: [revoked]
                    properties:
                      revoked:
                        type: boolean
                        const: true
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Endpoint not found or already revoked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /admin/upstream-r2/bulk-revoke:
    post:
      operationId: bulkRevokeUpstreamR2
      summary: Bulk soft-revoke multiple upstream R2 endpoints
      description: >
        Soft-revoke multiple upstream R2 endpoints in one request. Requires `confirm_count` to match
        the number of IDs as a fat-finger guard. Set `dry_run: true` to preview without executing.
      tags: [UpstreamR2]
      security:
        - CloudflareAccess: []
        - AdminKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids, confirm_count]
              properties:
                ids:
                  type: array
                  items:
                    type: string
                  minItems: 1
                  maxItems: 100
                  description: Array of upstream R2 endpoint IDs to revoke.
                confirm_count:
                  type: integer
                  description: Must exactly match the length of `ids`.
                dry_run:
                  type: boolean
                  default: false
                  description: Preview the operation without executing.
      responses:
        '200':
          description: Bulk revoke result or dry-run preview.
          content:
            application/json:
              schema:
                type: object
                required: [success, result]
                properties:
                  success:
                    type: boolean
                    const: true
                  result:
                    oneOf:
                      - $ref: '#/components/schemas/BulkResult'
                      - $ref: '#/components/schemas/BulkDryRunResult'
        '400':
          description: Validation error (empty array, confirm_count mismatch, over 100 items).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /admin/upstream-r2/bulk-delete:
    post:
      operationId: bulkDeleteUpstreamR2
      summary: Bulk permanently delete multiple upstream R2 endpoints
      description: >
        Permanently remove multiple upstream R2 endpoints from the database. Works on endpoints
        in any state. Requires `confirm_count` as a fat-finger guard. Set `dry_run: true` to preview.
      tags: [UpstreamR2]
      security:
        - CloudflareAccess: []
        - AdminKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids, confirm_count]
              properties:
                ids:
                  type: array
                  items:
                    type: string
                  minItems: 1
                  maxItems: 100
                  description: Array of upstream R2 endpoint IDs to delete.
                confirm_count:
                  type: integer
                  description: Must exactly match the length of `ids`.
                dry_run:
                  type: boolean
                  default: false
                  description: Preview the operation without executing.
      responses:
        '200':
          description: Bulk delete result or dry-run preview.
          content:
            application/json:
              schema:
                type: object
                required: [success, result]
                properties:
                  success:
                    type: boolean
                    const: true
                  result:
                    oneOf:
                      - $ref: '#/components/schemas/BulkResult'
                      - $ref: '#/components/schemas/BulkDryRunResult'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  # ─── Admin: Config ───────────────────────────────────────────────────────
...